rest_server_requirement.txt
flask
jsonpickle
Pillow